[~] Memoize calls
