+++ src/components/Tables/Jobs/JobsGridView.react.js
@@ -85,7 +85,7 @@ const toGrid = dataSource =>
 
 const JobsGridView = () => {
   const { dataSource } = useJobs();
-  return <>{toGrid(dataSource)}</>;
+  return <>{toGrid(dataSource.slice(0, 5))}</>;
 };
 
 export default memo(JobsGridView);
