+++ src/hooks/index.js
@@ -13,3 +13,4 @@ export { default as useTraceData } from 'hooks/useTraceData';
 export { default as useLocalStorage } from 'hooks/useLocalStorage';
 export { default as useDrawer } from 'hooks/useDrawer';
 export { default as useViewType } from 'hooks/useViewType';
+export { default as useLogs } from 'hooks/useLogs';
