+++ src/hooks/useNodeInfo.js
@@ -1,37 +1,13 @@
-import { useState, useEffect } from 'react';
-import useActions from './useActions';
+import { useState, useEffect, useMemo, useCallback } from 'react';
+import useLogs from './useLogs';
+import { nodeFinder } from 'utils';
 
 const EMPTY_NODE = {};
 
-const findNodeName = nodeName => node => node.nodeName === nodeName;
-
-const nodeFinder = ({ graph, pipeline }) => nodeName => {
-  const nodeData = graph.nodes.find(findNodeName(nodeName));
-  const node = pipeline.nodes.find(findNodeName(nodeName));
-  const { jobId } = pipeline;
-
-  const taskId =
-    nodeData && nodeData.taskId ? nodeData.taskId : nodeData.batch && nodeData.batch[0].taskId;
-  const podName =
-    nodeData && nodeData.podName ? nodeData.podName : nodeData.batch && nodeData.batch[0].podName;
-
-  const payload = {
-    ...nodeData,
-    jobId,
-    taskId,
-    nodeName,
-    podName,
-    origInput: node.input,
-    batch: nodeData.batch || [],
-  };
-
-  return payload;
-};
-
 const useNodeInfo = ({ graph, pipeline }) => {
   const [node, setNode] = useState(EMPTY_NODE);
-  const { getKubernetesLogsData: getLogs } = useActions();
-  const findNodeByName = nodeFinder({ graph, pipeline });
+  const { getLogs } = useLogs();
+  const findNodeByName = useCallback(nodeFinder({ graph, pipeline }), []);
 
   useEffect(() => {
     if (pipeline && pipeline.nodes) {
@@ -42,20 +18,23 @@ const useNodeInfo = ({ graph, pipeline }) => {
     }
   }, [findNodeByName, pipeline]);
 
-  const events = {
-    click: ({ nodes }) => {
-      const [nodeName] = nodes;
-      if (!nodeName) {
-        return;
-      }
+  const events = useMemo(
+    () => ({
+      click: ({ nodes }) => {
+        const [nodeName] = nodes;
+        if (!nodeName) {
+          return;
+        }
 
-      const payload = findNodeByName(nodeName);
-      const { taskId, podName } = payload;
+        const payload = findNodeByName(nodeName);
+        const { taskId, podName } = payload;
 
-      getLogs({ taskId, podName });
-      setNode(payload);
-    },
-  };
+        getLogs({ taskId, podName });
+        setNode(payload);
+      },
+    }),
+    [getLogs, findNodeByName],
+  );
 
   return { node, events };
 };
