+++ src/utils/index.js
@@ -5,4 +5,4 @@ export { default as transformTraceData } from './jaeger/transformTraceData';
 export { default as mapObjValues } from './mapObjValues';
 export { default as tryParse } from './handleParsing';
 export { default as notification } from './notification';
-export { formatNode, formatEdge, findNodeName } from './graph';
+export { formatNode, formatEdge, findNodeName, nodeFinder } from './graph';
